Rework DX compiler note machinery.
There were a lot of things wrong with the current setup and tests.

The dx notes test has been marked expected failure for forever, and
the conditions for when notes would be emitted was extremely
ad-hoc.

Under this new scheme, notes are almost never emitted. The only time a
note is emitted is when the note is actually user actionable, like
when an array with constant dimensions could be stack allocated but
isn't because it might overflow the stack. And the note actually
becomes helpful, telling you why it can't allocate it and what to
do. Otherwise, no one cares about what are essentially compiler
limitations. Also, we had a bunch of conditions about not emitting
notes for it doesn't make sense for the compiler to stack allocate
them, but how is a user supposed to know what makes sense to stack
allocate or not? Declaring something dynamic extent doesn't imply an
implementation strategy, strictly speaking. It just annotates a
property of the code, so that's why we didn't want to emit notes when
the user declares a constant dynamic extent or a function which isn't
a closure. Also of course emitting notes for deleted code was
undesirable, since this unactionable.

tl;dr

Correct DYNAMIC-EXTENT declarations should never be removed because a
user sees a note for something where stack allocation doesn't even
make sense, or the compiler simply can't do it.