Make sure to check arg types even during IR1 conversion in some cases.
f17002823fe2a114afcf04c4c3aa1055bace9035 made *ctype-test-fun* use NIL
as a sentinel. That actually changed the behavior of check-arg-types,
which for example should produce a note when we have unreachable code
in the arguments. So restore the noting behavior of the function and
write a regression test for it.

Makes progress towards lp#2016417, which doesn't comprehend the &rest
argument type correctly and has issues with the source transform.

There are also some other cases with defined types that this change
doesn't solve.