Revise allocation profiler patching
Calling (PATCH-ALL-CODE T) before saving a core will allow the profiler to
run without having to JIT-patch instructions later. This seems to be required
for certain sandboxing environments that refuse to map text pages as writable,
even on x86-64 linux where you'd think it's allowed. This is mainly an issue
for ELFinated cores. Otherwise all Lisp memory is mapped RWX.