Get #+(and immobile-space (not compact-instance-header)) working
By request of Charles Zhang the capabilities of immobile-space are
getting a slight refactor so that you can have a configuration which
demands immobile text but not immobile fdefns.

As for x86-64, there's no convenient way to request nonstandard
choices of build options because --without-compact-instance-header is
overridden by (WHEN (TARGET-FEATUREP :IMMOBILE-SPACE) in src/cold/shared.

If nothing else, this removes the restriction about trying to build
with arenas and no immobile-space, though the unit tests will fail.