Update to Unicode 11.0.0
Other than new defined characters, the most significant change is
revisions to the grapheme, word and line-breaking algorithms,
particularly as relates to emoji.

Grapheme-breaking needs a new mini-state-machine in order to track the
extended pictographic state.

Word-breaking changes just enough to invalidate our previous
strategy of (attempting to) implement Replacing Ignore Rules (section
6.2 of UAX 29); instead, we now look back when we are in an extended
context to find the correct class.  This is not maximally efficient,
and the strategy in line-breaking is probably better, but it is not
terrible either.  We also need to handle the new requirement to
consider non-Glue Zs characters as their own WSegSpace class.

The line-breaking needs a very minor adjustment: nothing breaks after
a zero-width joiner.