Support for newline variants in an external format
It might be nice to generate newline variants in external formats
on-demand, composing byte-to-byte and byte-to-character
transformations expressed at a high level, but with zero-cost
abstractions.  It might be nice.  But for now, do things the hard way:
generate all-new functions from all-new forms, and glue the explicit
external format object that that generates into
sb-impl::*external-formats*.