Support unibyte encodings with bytes without a character mapping
Most unibyte encodings have 256 characters, but not all: some have
bytes which map to no character.  For those, make sure that
DEFINE-UNIBYTE-EXTERNAL-FORMAT-WITH-NEWLINE-VARIANTS includes checking
code and a non-local exit to signal a decoding error, and that the
functions defined by DEFINE-UNIBYTE-TO-STRING-FUNCTIONS likewise check
the mapper for a non-code.

This allows us to express the ASCII external-format and its newline
variants using the generic definers.