Lift immobile space sub-2GB restriction for non-x86-64.
Also change arm64 Linux to use high address space unconditionally, to
match the behavior on other platforms and to show that immobile space
works fine above 4GB on this platform.

This also makes it easier to make immobile space work on Darwin and
NetBSD.

We need to disable a test that only succeeds half the time
though. It's way too flaky.