Fix FILE-POSITION in combination with UNREAD-CHAR
Since we can no longer reliably assure ourselves about the size of a
character in octets, it is no longer sufficient when reading a single
character to simply put the character at the end of the fd-stream's
input buffer: although that is correct for that state, a subsequent
UNREAD-CHAR will move the character buffer pointer back one space,
effectively making the last entry in the character size buffer live.

This means that when reading a single character, we must propagate not
only the element read, but also the number of bytes corresponding to
that character.  Alter the IN-STREAM convention to return the size as
a secondary value, including the size from any replacement characters.

We do not attempt to preserve the size information from a decoding
error leading to replacement, because it should not be the case that
an UNREAD-CHAR conceptually unwinds the stream: unreading the
replacement character will not trigger another decoding error when
READ-CHAR is next invoked; effectively, we do not unwind the file
position to before the decoding error was first encountered.  However,
in the absence of decoding errors, the FILE-POSITION reported after
READ-CHAR followed by UNREAD-CHAR should be the same as the original
FILE-POSITION.  Add tests for this to external-format tests with
invalid inputs.