Fix FILE-STRING-LENGTH given encoding errors
Make the external-format size function check the encodability of the
character before returning a size for it; this usually involves
calling a function that we already have.  In a few cases this may lead
to a duplicate call to CHAR-CODE, which is cheap but elidable if
necessary, though this function is only called in FILE-STRING-LENGTH
in any case.

(That's not quite true: it's also called in :UNREAD in fd-stream.lisp,
which is shadowed for FD-STREAMs with buffers, but exposed by
generating an FD-STREAM without a buffer by using :direction :io or
bivalent streams.)