Pass struct thread to arch_{set|clear}_pseudo_atomic
I initially was going to change these to take the context and thread pointers,
but no arch requires a context any more. Some previously used a bit in a CPU
register to indicate pseudo-atomic, but now it's always conveyed by bits in
a thread structure or static symbol value as per pseudo-atomic.h.
Except for ARM these could really be folded back into a single variant.

Remove obsolete comments about 'foreign_function_call_active'. The reasoning
was that if the P-A state was indicated by a regiser, then it would be correct
only if in Lisp code. As the comment says, a problem occurs if lisp enters
P-A and calls C or Lisp calls C which enters P-A. Deciding the nature of the
register would entail a stack walk.