Better unoptimized CEILING and FLOOR.
Don't transform everything to
(m-v-b (q r) (truncate x y) (if ... (+ q ...) (- q ..)))

because when X and Y are not known all these comparisons and
arithmetic operations will be full calls, making the generated code
very large and slow.
Transform only when TRUNCATE can operate on these numbers, meaning
word-sized integers and floats.
The rest goes to a new CEILING function, which performs the same
dispatch as TRUNCATE, avoiding multiple dispatches in TRUNCATE and
then in the generic arithmetic functions.