Eradicate assumption that weak-vector = SIMPLE-VECTOR internally
This change anticipates an incremental or concurrent GC wherein access to
weak objects has to synchronize with GC. There is an obvious race between
a collector deciding that an object was not strongly reachable (and
smashing it), and a mutator strengthening a weak reference merely by loading
it into a register.
Techniques to deal with the race typically involve a read barrier that may
involve fine-grained locking around weakRef.get(), or worst-case a way to
inhibit any weak object from being accessed while GC is deciding what to clear.
It would be highly suboptimal to put such a read barrier into SVREF.
To find problems, this change introduces a #+weak-vector-readbarrier feature
which is disabled by default. If enabled, SVREF won't work on the result
of MAKE-WEAK-VECTOR. xof notes that few users have taken the liberty of using
SVREF on SBCL's weak vectors. The migration path isn't too bad - it's just
a matter of changing SVREF to WEAK-VECTOR-REF.