Never warn when a type is disjoint from a previous derived type.
This warning never fired when *check-consistency* was off, but the
warning is actually not indicative of any real problem for many
legitimate cases. For example,

(defun f (x) (declare (unsigned-byte x) (float x)))

would cause this to warn because the lvar receiving the value of
(float x), after transformation, will receive (if (floatp x) x
...). The X in the then branch has type unsigned-byte and will
conflict with the lvar's type of FLOAT. Of course, deriving the type
of the REF node of X to be NIL in this case is completely legitimate:
this is consistent with the fact that the then branch of the if is
unreachable.