Reimplement :REPLACEMENT external-formats
Don't wrap a handler around every stream or octet function.  Instead,
thread the replacement from the external format through the various
functions all the way to the coding error site, and inline handling a
coding error if a replacement is provided.

In the process, also somewhat unify the behaviour of restarts around
stream and octet encoding- and decoding-errors: provide a USE-VALUE
restart whose argument should be a character, a string, an octet or an
octet sequence, which will be used as a replacement (with encoding or
decoding if necessary) to the operation that caused the error.

Note a couple of TODOs in the process.