check-consistency: Don't barf if an lvar has no uses.
Local mv-combinations can have lvars with no uses if
SUBSTITUTE-SINGLE-USE-LVAR kicks in and takes away the lvar's
use. This is surprisingly hard to fix at the source, since
mv-combinations do not have NIL stand in for flushed args, and it's
unclear how to assign consistent semantics to that without
complicating other parts of the IR.

Just disable the check for now, since having LVARs with a dest but no
use does not in itself necessarily mean the IR is inconsistent.