Turn on IR1 consistency checking for self-build.
Now we are able to run the full suite of IR1 consistency checks by
default for self-build, which is the first time this code has ever
been used to check SBCL self-build. (Even CMU CL can't completely
rebuild itself with consistency checking on currently, maybe since
1991 or so.)

On #+sb-devel we go even further and keep it on in the final image to
catch problems.

Doesn't work on arm64 because of issues described in parent commit.