Better handling of 2- and 4-byte external formats
Incrementing the buffer head by 1 (as in the previous commit) is in
general not a good strategy for 2- and 4-byte external formats.
Propagate the unit size of the external format to
stream-decoding-error-and-handle and to the resync-funtion, so that
restarts can advance by an appropriate amount.

In the process, notice various bits of breakage in octet routines for
2- and 4-byte external formats, where the routines assumed that the
subsequence contained an integer number of units, and read off the end
of the provided memory area.  Add checks for this case, and unify the
error signalling behaviour with that in fd-streams.

Add another clump of tests for these things.