Fix a file descriptor leak
Conditions signalled by OPEN and its callees after the file descriptor
is opened by the underlying operating system but before the stream
object is constructed and its finalizer installed would cause a file
descriptor leak.  Check explicitly for valid external formats at the
head of OPEN (and again at the head of MAKE-FD-STREAM, thanks to its
common use within the system); if a problem is detected in
MAKE-FD-STREAM, wrap an unwind-protect around signalling the error in
order to make sure that the file descriptor is closed on abort.