Remove early return
The if/else chain was supposed to land at the restoration of errno in all
cases. I believe none of the code above the early return would have messed up
errno, so returning was probably fine but definitely not needed.

Also, clang-tidy didn't like the indentation and I don't care to fight it.