Simplify xreffing inline expanded functionals.
No need to do all the extra stuff introduced in
30e65b004ace56e530469a364c35a6f5f5d686eb --
IR1-CONVERT-INLINE-EXPANSION already conveniently records the source
names in the right place, and optional dispatches are stored in their
entry points, so just use that. The actual global var lookup can be
done via the FREE-FUNS table, so we don't need to pun the
INLINE-EXPANDED slot.

Now IR1-CONVERT-INLINE-EXPANSION and IR1-CONVERT-LAMBDA-FOR-DEFUN are
even more similar, and the INLINE-EXPANDED slot on functionals has a
stronger type constraint.

The let conversion policy check doesn't even need to check
INLINE-EXPANDED; that was just a proxy for the fact that we weren't
checking the optional dispatch correctly for the normal INLINEP query.