Fix more asan complaints
It thinks calloc() in switch_to_arena leaks memory. I doubt it, because there
is a matching call to free() in free_thread_struct, but threads are freed
lazily. Considering that the arena save area was disabled in bf57b67b47
due to not working, don't allocate anything at all.