Lift forwarding-ptr.h and pointer_survived_gc_yet() into gencgc-impl.h
The mark-region collector wants a different function definition which it
injected into gc-common.c using #ifdefs when really it should be in the impl
header. Moreover, it declared pointer_survived_gc_yet as "extern" despite
fullcgc.c also definining a static function of that name, that the new impl
header couldn't actually be included without conflict.

Delete cheyneygc part of forwarding-ptr.h while we're at it.