stack: Remove NLE-BLOCK-FOO accessors.
These are done better by just checking the BLOCK-START-CLEANUP,
instead of doing pattern matching assuming that the REF of the first
node is to a constant leaf representing an NLX-INFO structure. The
cleanup is what we want to extract in all cases anyway.

NLE is also a misleading name -- TARGET is the terminology used by
NLX-INFO for the block containing the %NLX-ENTRY funny function.