ir2tran: Convert entry nodes more efficiently.
We are trying to generate code for each NLX-Info structure, so instead
of iterating through exits and checking for NLX-Info uniqueness by
consing a list and checking the cleanup kind, just check the cleanup
kind first and iterate directly over the NLX-Info structures. Also
update the comment to reflect this.