Allow conditional stack allocation even when not all branches can.
Due to the new stack pointer setting logic introduced in
4ce3711bdf52ab80bebf4b29f9c1c3f5f480124b, it is now possible to
allocate on the stack conditionally only from some of the branches of
a conditional instead of requiring that all branches be stack
allocatable. Previous to the mentioned change, doing so would've
caused the stack to be in an inconsistent state due to stack analysis
not being able to handle it and only a few IR2 translators actually
having logic to set the stack pointer only when it was possible to
stack allocate there as well.