Don't CLEAR-IR1-INFO if CLEAR-IR1-NAMESPACE already nukes everything.
The point of CLEAR-IR1-INFO is to remove left-over references to a
component after it has been compiled already, so that it can be
GC'd. But in this case the components are live until all infos are
cleared, and we clear the remaining tables referencing the leaves
which reference the component, essentially achieving the same effect,
but without deleting from the same list many times. CLEAR-IR1-INFO is
actually pretty slow when there are many components in the compilation
unit, as can happen with large block compilations, for example. Not
doing this allows a 15% speed-up when block-compiling node.lisp, which
is done during self-build.