Don't export GET/SET/END??ENT or WITH-*-DATABASE macros
It's a little hard to envision a use for the GET/SET/END??END routines
that the DO- macros don't cover. Since the GET/SET/END routines are
not reentrant, calling them during a DO- iteration can only serve to
mess up that iteration, so making those functions available seems only
to expose error possibilities. And if those functions are not external
to SB-POSIX, then users don't need WITH-*-DATABASE macros either.