stack: Make sure to not clean up nested stack values too early.
This was a regression that happened when we started only tracking
stack lvars instead of all dynamic extent values on the stack. It
turns out that the invariant that stack objects with different
lifetimes isn't true, even without SETQ, as the first test case
shows. Fix this by adding a funny function which only serves to keep
the stack fixed when dynamic extent values with different lifetimes
get pushed in an interleaved manner onto the stack. The invariant that
multiple-values and dynamic-extent lifetimes cannot be interleaved is
still true however.

Also add a test case where SETQ caused the same problem for good
measure.