stack: Preserve lvars under stack consed objects in a different way.
Do this directly with a PRESERVE-INFO slot on dynamic extent which
tracks which lvars need to be preserved. This is more robust that
trying to have a funny function do the work for us. We make liveness
analysis enliven stuff that needs to be preserved again but doing so
based on control flow, in a way that still allows us to oinly track
stack lvars in PUSHED and POPPED sets (keeping conditional stack
allocation much simpler).

Fixes a bug reported by Andreas Franke.