Improve conformance with CLHS 7.6.5
Prompted by a report by Daniel Kochmański: rework the keyword argument
checking in effective methods.  The checking needs to be inserted by
the caller of compute-effective-method; it should be orthogonal to the
effective method computation.  The fact that SBCL elides the checking in
the dispatch function, instead relying on the effective method
function to perform the check, is an implementation detail, and the
presence of the local pseudo-macro in the standard effective method
causes problems to particular custom uses of the MOP, for example the
very simple implementation of COMPUTE-DISCRIMINATING-FUNCTION in the
new test file.

As a (very welcome) side effect, this fixes the fact that we
apparently did not perform keyword applicability checking if there
were applicable auxiliary methods on the generic function.