Eliminate the magic in COMPUTE-EFFECTIVE-METHOD
Rather than have variously magical conditions to signal to callers
that there should be some action to take, perform an explicit check
for

Now the various COMPUTE-EFFECTIVE-METHODs are essentially pure, with
the exception that if the applicable methods have no primary methods,
or have a method with an invalid qualifier, they return forms
containing undumpable objects.  This matters at least for precom2.lisp
in the PCL bootstrap, and possibly other users of precompilation.