x86-64: Simplify global closure trampolines
The 6-word hand-assembled code blob hack is gone.
Now that every funcallable-instance has embedded code, the special case for
closures installed as globally named functions is not needed.
Each closure still needs a fresh wrapper, but the wrapper is a funcallable
instance that calls a closure as its implementation function.

Moreover, WITH-PINNED-OBJECTS need not pin FDEFN when setting it, but this
change could have been done long ago. FDEFNs used to store a rel32 branch
to the target, so both the fdefn and target had to be pinned (though in fact
both are immobile so I'm not sure why the pedantry). However, after git rev
1f892df960 (ca. 2019) FDEFN-RAW-ADDRESS is an absolute address.