Disable *COMPILE-VERBOSE* and *LOAD-VERBOSE* for --script
Scripts using COMPILE or LOAD, either directly or through third party systems,
pollute the standard output with messages unrelated to the script itself.

This is undesirable (to quote a comment in toplevel.lisp, "scripts don't need
to be stylish or fast, but silence is usually a desirable quality") and can
cause bugs when the output of a script is processed through a UNIX pipe by a
program expecting to get output and not verbose diagnostic messages.

Problem originally signaled by Hraban Luyat <hraban@0brg.net>.