x86: fix NetBSD build
Stores to static space which alias to a dynamic space card must not store a
mark, otherwise the next memory fault on the spuriously marked dynamic space
card will fail the check that an already-dirtied card should not fault.
This problem was not NetBSD specific, but exposed due to different starting
addreses for static and dynamic space than on Linux. In fact I think the bug
exists on all arch/OS combinations that use a mixture of logical and MMU-based
marking (e.g. including darwin-jit + soft-marks) so we should really fix them
all because apparently the code was getting lucky pretty much everywhere.