Add #define to watch all mmap/mprotect calls
Piggyback on the pre-existing layer of indirection (sbcl_mmap -> mmap)
which works around non-thread-safe libc interceptors of mmap.
It's not possible to both enable tracing and use the direct-syscall,
because the tracer uses the ordinary libc entry points, but that's ok.
It should be easier now to discover why mark-region on macOS fails.