Remove gencgc-release-granularity
and also give remap_page_range a better name to express intent,
which is decommit (release any backing storage for) the range.
POSIX madvise can't do that because the DONTNEED option is not the
same thing, except on Linux where it is but only to the legacy
madvise and not posix_madvise.
On #+win32 use the API directly.