mark-region: add minimal compact-instance-header feature
To get ths feature you have to add --with-permgen which is like an extension
of static space for layouts only. I plan to make the space relocatable and
potentially GCable after figuring out how much mutator performance loss
without compact-instance-header is due to greater memory consumption, and/or
whether it's worth the trouble to deal with mortality or lack thereof.