Remove eql-hash-no-memoize
It merely avoided a writeback of symbol-hash in GETHASH, the theory being
that it's OK to read a hash of 0 which would result in a miss (probably)
which was fine consdering that any symbol present in the table would have
had its hash computed by PUTHASH if not before.
But every symbol has a hash computed by MAKE-SYMBOL now.