Be stricter about specifying hash functions for CTYPE slots
Remove FALLBACK-SXHASH. Instead choose from NUMBER-HASH (which has no
obligation to match SB-IMPL::NUMBER-SXHASH) or FALLBACk-HASH for HAIRY type,
or no hash at all for SIMD-PACK tag which is essentially already hashed.
The intent is to clarify where the cross-compiler must match the target's
SXHASH, particularly in regard to symbols, and whether pseudo-random hashes
are or are not permitted.