Express ctype->hashset-sym differently for make-host-2
I started to improve TYPECASE on structure types that are not sealed
(so subtypes could be added) using a PCL cache. The cross-compiler wanted to
use that transform on this function, but it failed in cold-init (after making
caches themselves viable, of course). NOTINLINE TYPEP made things worse,
because parsing would have work sooner. This change avoids relying on an
abstract rendering of TYPEP by making it concrete albeit ad-hoc.