Only COERCE-TO-FUN if the result-type is exactly FUNCTION
It's definitely wrong to attempt to coerce symbols and lambda forms to
subclasses of FUNCALLABLE-STANDARD-OBJECT.  (This move to "exactly
FUNCTION" might be too strong, though I think it's probably a bit
weird to expect COERCE to do anything sensible with contorted subtypes
of FUNCTION, e.g. (COERCE '+ '(AND FUNCTION (SATISFIES HALTS-P))).