Quiet a clisp warning
By CLHS 6.1.1.4, it is legal for an implementation to provide an
initialization form in the loop prologue for any iteration variable.
That means that in principle using a name for a loop variable that
shadows an external name, and expecting to be able to refer to that
external name during any part of the loop body, is unportable.

(In practice I don't think that clisp, or any other implementation,
actually does the inconvenient thing: but it's easy enough to choose
some different variable names.)