Fix a cross-compiler portability issue
We simplify type tests of union types by examining whether they
contain within their types any sets of types that can be tested using
a backend type predicate.  If it does, then we test for that type,
then for the original type with those types removed.

It's moderately natural to express that removal with set-difference --
particularly after the types began being hash-consed.  However, using
set-difference introduces a dependence on the host Lisp's
implementation, as the order of the elements of the returned
list (interpreted as a set) is not specified.  Using remove-if with an
explicit membership test preserves the original order of types.

This was tracked down by noticing:
- the cross-compiled lisp-obj file for src/code/type-class.lisp had
  different contents, the first difference being a dumped word-sized
  integer;
- that integer corresponded to the encoded-locs field of a
  compiled-debug-fun associated with !allocate-numeric-type
- the code size for !allocate-numeric-type was different depending on
  the host compiler;
- the trace file for the cross-compiled src/code/type-class.lisp
  revealed that the type testing for LOW (and later HIGH) was done in
  a different order.