Speed up NAME-CHAR and shrink the lookup tables
We can exploit the fact that Huffman decoding is much faster than encoding.
Perfectly map a name to 0..N and then get the codepoint at index N. If that
alleged codepoint inverts (using CHAR-NAME) to the name as supplied, then we
found it; but care must to be taken with characters that have multiple names.

On top of the non-consing and >5x speedup (on x86-64), we get a core size
reduction of 1.5% from this; the space savings for 32-bit is negligible.