Style-warn if TYPECASE has an interior T clause
Additionally, if [EC]TYPECASE contains a T clause then expand without
an error call. The compiler would discover that the error can't occur,
but there's no reason to put it in. (Lots of people use ETYPECASE
with a T clause) The related scenario involving [EC]CASE can't happen
because T is not allowed except as the final clause.