Sort the package-use-lists in genesis
CL's PACKAGE-USE-LIST is conceptually a set, and implementations
are (empirically) allowed to return the packages in any order.  Remove
the sensitivity of cold-sbcl.core to this host implementation decision
by sorting the package names in the use list before adding to
our *package-graph*.

After this change, cold-sbcl.core as produced by cmucl and sbcl are
identical.  There remains some non-determinism in the post-cold-init
build.