Add ldb command for hash-table checker
The C checker unfortunately produces false positives. I tried using a big
hammer of WITHOUT-GCING around REMHASH and CLRHASH, which seemed to hide
actual problems. That's not good. This change should help debug the checker
so we can use it to debug the GC. And I think the pmrgc still has issues
with not setting need-to-rehash often enough, hence the need for this.