Fix check-deadlock.
Threads may relinquish and acquire new locks while check-deadlock is
trying to find a thread that either doesn't wait for anything or waits
for the current thread's lock. Which creates two problems
a) it may never terminate (or take a lot of cycles)
b) it may produce false positives.

Limit the depth to 10 threads.
Recheck that the previous mutex is still being held by the same thread
when a cycle is found.

Reported by Michał Herda.