Break the build on float cache mismatch
Now that logarithms are correctly computed (looked up) on the host,
all of the values computed should match the target system even for
older host SBCLs (where the native computation of (log 2d0 10) was off
by 1ulp).  We can now break the build, rather than just printing, if
something computed during cross-compilation does not match the
target's floating point computation.

In principle this is still problematic for two reasons.

The less fundamental one is that we have not yet finished implementing
portable versions of all float routines used; there are still some
calls to REALNUMIFY and related functions, which happen to give the
right answer in all currently known cross-compilation hosts.
Finishing the implementation should be straightforward, with the
implementation of a few more oracles and some rational arithmetic.

The more fundamental problem is that our production implementation of
transcendental functions delegates to the operating system libraries,
and there is no guarantee that the values that we compute are actually
those produced by the operating system in all possible configurations
at all possible times.  It is theoretically possible, say, that an
upgrade of the C library could bring in a new libm with different
results: or a user could be building for a replacement math library
with different errors.  (This isn't just a problem in
cross-compilation; the potential time-varying results of pure math
routines is problematic for compilation / runtime consistency in
production too, for example regarding type derivation).