Make stuff regarding debug names much less complex.
Don't show the full binding form for LET function debug names, just
show the names of the variables bound. This makes it so that
arbitrarily nested hairy stuff is NEVER debug-namified. This means
instead of truncating debug name forms, we can just use them as-is,
avoiding consing new objects. No new stuff is dumped that wasn't
already going to get dumped anyway. Printer control variables can be
used to display the debug names differently if need be, we don't need
to reinvent that. It was also found that we were passing literal leaf
objects to the &aux debug name, so don't do that.

The upshot is that we can get rid of all the debug-name-marker
structures. There's no reason that something that exists purely for
compiler debugging should be privileged and have its own fop.