Fix integer rounders' interaction with minus zero
In #+round-float builds, we had been getting the appropriately-rounded
integer value as a float (in order to compute the remainder) directly
from float rounding, rather than converting the rounded integer back
to a float.  This was correct for all results except when the rounded
float was minus zero, which could lead to minus zero being lost in the
remainder.  Subtracting minus zero from the rounded float makes this
correct in all cases.