debug-info: Use a weak side table for COMPILED-DEBUG-FUN caching.
Having a separate slot doesn't buy much other than waste about 200kb
of space in a warm core. The debugger doesn't need accesses to the
cache to be fast.

This also fixes a bug in editcore, allowing us to remove a bunch of
allowed instances that never get dumped as part of COMPILED-DEBUG-INFO
now.