Remove sb-editcore::%fun-name-from-core
It's subsumed by extract-object-from-core. Confusingly some of the "junk"
in names was removed by %fun-name-from-core while other junk was removed in
remove-name-junk. Now it's all cleaned up in the latter.

The new logic is not at all concerned with whether a Lisp symbol is external
or internal in its package, because frankly it's not important.
A typical ELFinated application will have thousands upon thousands of C symbols
spelled "lambda_N" so it scarcely matters whether  actually-distinct names
have 1 or 2 colons in them.