Eliminate a perfect hash function from fixup application
Fixup flavors are numerically encoded in fasl dumping, so it is better to use
a CASE over the encodings rather than decode to a symbol and then perfectly
hash symbols that correspond to pre-determined small integers.

This fixes failure-to-build on non-SBCL with #+(and x86-64 (not sb-thread)),
in particular dragonfly BSD, which prior to rev ff7cc435 was also missing
a perfect hash for static symbols.