Just use DEFPACKAGE to define our packages.
Instead of using a not-quite-the-same-yet-also-limited package
definition mechanism which differed in the order it did its operations
from DEFPACKAGE. This means we do our forward declarations explicitly
now to resolve forward references, which makes it clearer what gets
imported from where when. We also get :SHADOWING-IMPORT-FROM for free
now, not that we use it currently.

* Some imports/exports are fixed so that symbols get accessed via a
canonical package.

* Move a bunch of regalloc imports into pack.lisp to make it easier to
resolve the references without duplicating the symbols. There are so
many that I'm wondering whether or not it should all just be SB-C
again.

* Hack the reader (with extremely limited scope) to make "CL" <->
"XC-STRICT-CL" renaming come out right. This allows us to not have to
add special logic in chill to deal with the renaming and allows us to
have the package definitions oblivious to our bootstrap model.

* chill now warns about package variance. This is correct, since we
are reloading in package definitions that can have less exports than
the target image. If this gets annoying we could suppress the warning,
but it's helpful to see what differs in any case.