Improve NAME-CONFLICT for IMPORT
Collect all symbols of the same name from the designated list of
symbols to import, so that in the case of collisions between imported
symbols (and possibly already-accessible symbols) we can present the
user with a single such conflict to resolve (per name) rather than a
sequence of them.

Make NAME-CONFLICT for IMPORT take as a datum not a single symbol, but
all the symbols requested for import of the same name.  Any
already-accessible symbol should be passed in as an element of the
SYMBOLS &rest list.

Make the pairwise restarts SHADOWING-IMPORT-IT and DONT-IMPORT-IT only
active when there are two symbols involved in the conflict, one of
which is accesible.

A comment (but not much of the subsequent code) is invalidated by this
change, so delete it.  Instead, make sure that if there is no current
accessible symbol of the name being imported, the resulting operation
should be an IMPORT, not a SHADOWING-IMPORT.

This fixes long-standing test :IMPORT-SINGLE-CONFLICT in
packages.impure.lisp