sb-cover: Stop dumping +code-coverage-unmarked+.
There's no need to dump conses of the form (path
. +code-coverage-unmarked+) into the fasl file itself. Just dump the
paths and create the fresh conses in the loader. This avoids problems
with list coalescing so that we can just use NIL as the unmarked
marker now, allowing us to remove a lot of pointless abstractions. It
also shrinks the fasl size.

We can also now move a bunch of sb-cover stuff out of the compiler and
the compiler package into the contrib proper.