Better handling of mixed arguments within LOG
Each of the number and base arguments to two-argument LOG can be
rational, single, double, complex-rational, complex-single or
complex-double, meaning that there are thirty-six generic cases to
handle, even after some special cases have been picked off.  In
addition, the return type in some cases needs a little care because
not all complex results from real inputs have been excluded by the
time we get to the number-dispatch.

We need to take care over this in order to effectively do the floating
point contagion between the arguments to two-argument LOG before the
decomposition as the ratio of the logarithms of the arguments, as
otherwise the difference in precision leads to substantially
inaccurate final results.

Out-of-line LOG becomes somewhat gigantic as a result of this.  Make
the transforms for floating point LOG at least plausibly
efficient (and also make sure they do not lose precision either).

This fixes the long-standing (:LOG :SAME-BASE-DIFFERENT-PRECISION)
test failure (only recently added, but previously noted in a comment).